# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Saúde UNASP",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/cadastro/validar/documento/{documento}": {
            "get": {
                "tags": [
                    "Cadastro"
                ],
                "summary": "Validar documento",
                "description": "Valida se documento é válido para cadastro",
                "operationId": "706ca24c5d9369d129a03b17f5ec23bb",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    },
                    {
                        "name": "documento",
                        "in": "path",
                        "description": "CPF",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "cpfValido": {
                                "summary": "CPF Válido",
                                "value": "00000000191"
                            },
                            "cpfInvalido": {
                                "summary": "CPF Inválido",
                                "value": "00000000190"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Documento válido",
                                        "value": {
                                            "message": "Este documento é válido"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Documento é inválido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Documento Inválido",
                                        "value": {
                                            "message": "Este documento é inválido"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Documento já foi cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Documento Já Cadastrado",
                                        "value": {
                                            "message": "Este documento já foi cadastrado"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cadastros": {
            "get": {
                "tags": [
                    "Cadastro"
                ],
                "summary": "Listar cadastros",
                "description": "Lista todos os cadastros",
                "operationId": "245018355fc2e09b015fbb35c782c11d",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "ListaCadastro": {
                                        "summary": "ListaCadastro",
                                        "value": {
                                            "cadastros": [
                                                {
                                                    "id": 1,
                                                    "nomeCompleto": "João Silva",
                                                    "dataNascimento": "1920-01-01",
                                                    "sexo": "M",
                                                    "cep": "05555000",
                                                    "cpf": "12345678901",
                                                    "uf": "SP",
                                                    "cidade": "São Paulo",
                                                    "logradouro": "Rua do teste",
                                                    "numeroLogradouro": "202 A",
                                                    "email": "joao@teste.com",
                                                    "created_at": "2000-01-01T12:00:00.000000Z",
                                                    "updated_at": "2000-01-01T12:00:00.000000Z"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Não tem conteúdo"
                    }
                }
            }
        },
        "/api/v1/cadastro/{documento}": {
            "get": {
                "tags": [
                    "Cadastro"
                ],
                "summary": "Consultar cadastro pelo documento",
                "description": "Consultar cadastro pelo documento",
                "operationId": "81bd906c97113cd1f0b1905660e54883",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    },
                    {
                        "name": "documento",
                        "in": "path",
                        "description": "CPF",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "cpfValido": {
                                "summary": "Documento",
                                "value": "00000000191"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastroResult"
                                },
                                "examples": {
                                    "result": {
                                        "summary": "CadastroResult",
                                        "value": {
                                            "id": 1,
                                            "nomeCompleto": "João Silva",
                                            "dataNascimento": "1920-01-01",
                                            "sexo": "M",
                                            "cep": "05555000",
                                            "cpf": "12345678901",
                                            "uf": "SP",
                                            "cidade": "São Paulo",
                                            "logradouro": "Rua do teste",
                                            "numeroLogradouro": "202 A",
                                            "email": "joao@teste.com",
                                            "created_at": "2000-01-01T12:00:00.000000Z",
                                            "updated_at": "2000-01-01T12:00:00.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Documento não cadastrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cadastro"
                ],
                "summary": "Apaga cadastro pelo documento",
                "description": "Apaga cadastro pelo documento",
                "operationId": "d74a0d6d685f1033b3527d7015a4cd11",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    },
                    {
                        "name": "documento",
                        "in": "path",
                        "description": "CPF",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "cpfValido": {
                                "summary": "Documento",
                                "value": "00000000191"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cadastro apagado com sucesso"
                    },
                    "404": {
                        "description": "Documento não cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroResult"
                                },
                                "examples": {
                                    "ErroResult": {
                                        "summary": "ErroResult",
                                        "value": {
                                            "message": "Documento não cadastrado"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErroResult": {
                "title": "ErroResult",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CadastroResult": {
                "title": "CadastroResult",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nomeCompleto": {
                        "type": "string"
                    },
                    "dataNascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "sexo": {
                        "type": "string"
                    },
                    "cep": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "uf": {
                        "type": "string"
                    },
                    "cidade": {
                        "type": "string"
                    },
                    "logradouro": {
                        "type": "string"
                    },
                    "numeroLogradouro": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}